requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
requests>=2.31
gunicorn>=21.0
psycopg2-binary>=2.9
mysqlclient>=2.2
python-decouple>=3.8
django-cors-headers>=4.3
hypothesis>=6.90
pytest-django>=4.7
pytest>=7.0
